pythonProject/requirements.txt
pytest~=7.1.3
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest

PyYAML~=6.0
requests~=2.28.1